backend/admin_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
googletrans==4.0.0rc1
pydantic